Added feature to take screenshots from folders Added couple more utils methods

+++ src/utils/Zip_Folder.py
@@ -5,9 +5,10 @@ from utils.Misc import Misc
 
 
 class Zip_Folder():
-    def __init__(self, target_folder=None):
-        self.target_folder = target_folder
-        self.zip_file      = None
+    def __init__(self, target_folder=None,delete_zip_file=True):
+        self.target_folder   = target_folder
+        self.zip_file        = None
+        self.delete_zip_file = delete_zip_file
 
     def __enter__(self):
         if Files.exists(self.target_folder):
@@ -15,5 +16,5 @@ class Zip_Folder():
         return self.zip_file
 
     def __exit__(self, type, value, traceback):
-        if Files.exists(self.zip_file):
+        if Files.exists(self.zip_file) and self.delete_zip_file:
             Files.delete(self.zip_file)
